Refactor methods from LoginPage to HomePage
