+++ SeleniumTests/BaseTests/BaseTest.cs
@@ -1,6 +1,7 @@
 ﻿using NUnit.Framework;
 using NUnit.Framework.Interfaces;
 using SeleniumFramework;
+using SeleniumFramework.Pages.PiguLT;
 using SeleniumFramework.Pages.PiguLTPage;
 
 namespace SeleniumTests.BaseTests
@@ -17,9 +18,9 @@ namespace SeleniumTests.BaseTests
             // Login to page
             // If you get the error "Too many logins, please try later." when trying to login, change the "EnterLoginEmail" variable. Select one from four:
             // InputEmail1; InputEmail2; InputEmail3; InputEmail4;
-            LoginPage.ClickAllowAllCookies();
-            LoginPage.ClickOnProfileIcon();
-            LoginPage.ClickOnBlueLoginButton();
+            HomePage.ClickAllowAllCookies();
+            HomePage.ClickOnProfileIcon();
+            HomePage.ClickOnBlueLoginButton();
             LoginPage.EnterLoginEmail(TestData.Login.InputEmail1);
             LoginPage.EnterLoginPassword(TestData.Login.InputPassword);
             LoginPage.ClickOnGreenLoginButton();
